+++ apps/festival/festival/src/app/dashboard/tunnel/summary/summary.module.ts
@@ -5,8 +5,8 @@ import { RouterModule } from '@angular/router';
 import { TunnelPageModule } from '@blockframes/ui/tunnel';
 import { TunnelSummaryComponent } from './summary.component';
 
-// Summary components
-import { MovieTunnelSummaryModule } from '@blockframes/movie/pages/summary/summary.module';
+import { MovieTunnelSummaryModule } from '@blockframes/movie/form/summary/summary.module';
+import { GetPathModule } from '@blockframes/utils/pipes/get-path.pipe';
 
 // Material
 import { MatCardModule } from '@angular/material/card';
@@ -21,8 +21,8 @@ import { MatDividerModule } from '@angular/material/divider';
     CommonModule,
     FlexLayoutModule,
     TunnelPageModule,
-    // Summary components
     MovieTunnelSummaryModule,
+    GetPathModule,
     // Materials
     MatCardModule,
     MatIconModule,
