+++ libs/movie/src/lib/movie/form/summary/additional-information/additional-information.component.html
@@ -102,7 +102,7 @@
 <!-- CERTIFICATION -->
 <article>
   <h6>
-    Qualification{ movie.qualifications.value.length, plural, =0 {} =1 {} other {s}}
+    Qualification{ movie.certifications.value.length, plural, =0 {} =1 {} other {s}}
     <a mat-icon-button [routerLink]="link" fragment="qualifications"><mat-icon svgIcon="pencil" color="primary"></mat-icon></a>
   </h6>
   <ul>
