--- libs/movie/src/lib/movie/form/summary/summary.component.html
@@ -1,130 +1,0 @@
-<tunnel-page>
-  <ng-container tunnel-form>
-    <h4 title>LAST STEP</h4>
-    <h1 subtitle>Summary Page</h1>
-    <p description class="mat-body-1">
-      Please verify that all the information you have entered is correct before submitting to Archipel Market for validation.
-    </p>
-
-    <div *ngIf="missingFields.length || invalidFields.length" class="mat-body">
-      <span *ngIf="missingFields.length" class="missing-control">You have {{ missingFields.length }} missing fields below.</span>
-      <span *ngIf="invalidFields.length" class="mandatory-control">You have {{ invalidFields.length }} mandatory fields to fill in.</span>
-    </div>
-
-    <!-- MAIN INFORMATION -->
-    <section id="main-information">
-      <h3>Main Information</h3>
-      <article>
-        <movie-summary-main [movie]="form" [link]="getPath('main')"></movie-summary-main>
-      </article>
-    </section>
-
-    <mat-divider></mat-divider>
-
-    <!-- STORYLINE ELEMENTS -->
-
-    <section id="storyline-elements">
-      <h3>Storyline Elements</h3>
-      <article>
-        <movie-summary-synopsis [movie]="form" [link]="getPath('synopsis')"></movie-summary-synopsis>
-      </article>
-    </section>
-
-    <mat-divider></mat-divider>
-
-    <!-- PRODUCTION INFORMATION -->
-
-    <section id="production-information">
-      <h3>Production Information</h3>
-      <article>
-        <movie-summary-production [movie]="form" [link]="getPath('production')"></movie-summary-production>
-      </article>
-    </section>
-
-    <mat-divider></mat-divider>
-
-    <!-- ARTISTIC TEAM -->
-
-    <section id="artistic-team">
-      <h3>Artistic Team</h3>
-      <article>
-        <movie-summary-artistic [movie]="form" [link]="getPath('artistic')"></movie-summary-artistic>
-      </article>
-
-    </section>
-
-    <mat-divider></mat-divider>
-
-    <!-- SELECTION AND REVIEWS -->
-
-    <section id="reviews">
-      <h3>Selections and Reviews</h3>
-      <article>
-        <movie-summary-review [movie]="form" [link]="getPath('reviews')"></movie-summary-review>
-      </article>
-    </section>
-
-    <mat-divider></mat-divider>
-
-    <!-- ADDITIONAL INFORMATION -->
-
-    <section id="additional-information">
-      <h3>
-        Additional Information
-      </h3>
-      <article>
-        <movie-summary-additional-information [movie]="form" [link]="getPath('additional-information')"></movie-summary-additional-information>
-      </article>
-    </section>
-
-    <mat-divider></mat-divider>
-
-    <!-- TECHNICAL INFORMATION -->
-
-    <section id="technical-information">
-      <h3>Technical Information</h3>
-      <article>
-        <movie-summary-technical-info [movie]="form" [link]="getPath('technical-info')"></movie-summary-technical-info>
-      </article>
-
-    </section>
-
-    <mat-divider></mat-divider>
-
-    <!-- AVAILABLE MATERIALS -->
-
-    <section id="available-materials">
-      <h3>Available Materials</h3>
-      <article>
-        <movie-summary-available-materials [movie]="form" [link]="getPath('available-materials')"></movie-summary-available-materials>
-      </article>
-    </section>
-
-    <mat-divider></mat-divider>
-
-    <!-- PROMOTIONAL ELEMENTS -->
-
-    <section id="promotional-elements">
-      <h3>Promotional Elements</h3>
-      <article>
-        <movie-summary-media-files [movie]="form" [link]="getPath('media-files')"></movie-summary-media-files>
-      </article>
-
-      <article>
-        <movie-summary-media-images [movie]="form" [link]="getPath('media-images')"></movie-summary-media-images>
-      </article>
-
-      <article>
-        <movie-summary-media-videos [movie]="form" [link]="getPath('media-videos')"></movie-summary-media-videos>
-      </article>
-
-    </section >
-
-    <!-- Submit -->
-    <button test-id="submit" mat-flat-button color="accent" class="submit" (click)="submit()" [disabled]="isPublished$ | async">
-      <span>Publish to Marketplace</span>
-      <mat-icon svgIcon="arrow_forward"></mat-icon>
-    </button>
-  </ng-container>
-
-</tunnel-page>
